1. Refactoring 대상 : ConfusionGas, ParalyticGas, StenchGas 3개의 Class내의 함수 evolve, use, tileDesc 3개의 함수에 대하여 Refactoring 수행(모두 동일한 이름의 함수를 지님)
ConfusionGas, ParalyticGas, StenchGas, three functions in the class,
evolve, use, and tileDesc. Refactoring is performed for three functions
(all have the same name function).

2. 적용한 Refactoring Operation 이름
: Move Method

3. (1)의 Refactoring Operation을 적용한 이유
:3개의 함수가 모두 중복되는 코드를 가지기 때문에 새로운 Class를 만들어서 옮김

Since all three functions have redundant code, I created a new class and
moved it.
