+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/Gas.java
@@ -1,0 +1,43 @@
+package com.shatteredpixel.shatteredpixeldungeon.actors.blobs;
+
+import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
+import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
+import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
+import com.shatteredpixel.shatteredpixeldungeon.effects.BlobEmitter;
+import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
+import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
+
+public abstract class Gas extends Blob {
+	@Override
+	protected void evolve() {
+		super.evolve();
+		
+		Char ch;
+		int cell;
+		
+		for (int i = area.left; i < area.right; i++) {
+			for (int j = area.top; j < area.bottom; j++) {
+				cell = i + j * Dungeon.level.width();
+				if (cur[cell] > 0 && (ch = Actor.findChar(cell)) != null) {
+					if (!ch.isImmune(this.getClass()))
+						Gas_evolve(ch);
+				}
+			}
+		}
+	}
+	
+	@Override
+	public void use( BlobEmitter emitter ) {
+		super.use( emitter );
+
+		emitter.pour( Speck.factory(Speck.STENCH), 0.4f );
+	}
+
+	@Override
+	public String tileDesc() {
+		return Messages.get(this, "desc");
+	}
+
+	protected abstract void Gas_evolve(Char ch);
+	
+}
