+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/ParalyticGas.java
@@ -21,16 +21,12 @@
 
 package com.shatteredpixel.shatteredpixeldungeon.actors.blobs;
 
-import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
-import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Paralysis;
-import com.shatteredpixel.shatteredpixeldungeon.effects.BlobEmitter;
-import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
-import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
+import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Gas;
 
-public class ParalyticGas extends Blob {
+public class ParalyticGas extends Gas {
 	
 	{
 		//acts after mobs, to give them a chance to resist paralysis
@@ -38,32 +34,7 @@ public class ParalyticGas extends Blob {
 	}
 	
 	@Override
-	protected void evolve() {
-		super.evolve();
-		
-		Char ch;
-		int cell;
-
-		for (int i = area.left; i < area.right; i++) {
-			for (int j = area.top; j < area.bottom; j++) {
-				cell = i + j * Dungeon.level.width();
-				if (cur[cell] > 0 && (ch = Actor.findChar(cell)) != null) {
-					if (!ch.isImmune(this.getClass()))
-						Buff.prolong(ch, Paralysis.class, Paralysis.DURATION);
-				}
-			}
-		}
-	}
-	
-	@Override
-	public void use( BlobEmitter emitter ) {
-		super.use( emitter );
-		
-		emitter.pour( Speck.factory( Speck.PARALYSIS ), 0.4f );
-	}
-	
-	@Override
-	public String tileDesc() {
-		return Messages.get(this, "desc");
+	protected void Gas_evolve(Char ch) {
+		Buff.prolong( ch, Paralysis.class, Paralysis.DURATION);
 	}
 }
