+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/StenchGas.java
@@ -21,44 +21,14 @@
 
 package com.shatteredpixel.shatteredpixeldungeon.actors.blobs;
 
-import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
-import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Paralysis;
-import com.shatteredpixel.shatteredpixeldungeon.effects.BlobEmitter;
-import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
-import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
+import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Gas;
 
-public class StenchGas extends Blob {
-
+public class StenchGas extends Gas {
 	@Override
-	protected void evolve() {
-		super.evolve();
-
-		Char ch;
-		int cell;
-
-		for (int i = area.left; i < area.right; i++){
-			for (int j = area.top; j < area.bottom; j++){
-				cell = i + j*Dungeon.level.width();
-				if (cur[cell] > 0 && (ch = Actor.findChar( cell )) != null) {
-					if (!ch.isImmune(this.getClass()))
-						Buff.prolong( ch, Paralysis.class, Paralysis.DURATION/5 );
-				}
-			}
-		}
-	}
-
-	@Override
-	public void use( BlobEmitter emitter ) {
-		super.use( emitter );
-
-		emitter.pour( Speck.factory(Speck.STENCH), 0.4f );
-	}
-
-	@Override
-	public String tileDesc() {
-		return Messages.get(this, "desc");
+	protected void Gas_evolve(Char ch) {
+		Buff.prolong( ch, Paralysis.class, Paralysis.DURATION/5 );
 	}
 }
