+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/blobs/test/ParalyticGasTest.java
@@ -1,0 +1,29 @@
+package com.shatteredpixel.shatteredpixeldungeon.actors.blobs.test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class ParalyticGasTest {
+
+	@Test
+	void testGas_evolve() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testUse() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testEvolve() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testTileDesc() {
+		fail("Not yet implemented");
+	}
+
+}
